commit: separate ACRHyperLinkTextView, comman textproperties converter
